refactor(wasip2): declare shared WIT types in their own package
The datetime record was defined in wasi_fs.go, which is not where it
belongs: wall-clock owns `datetime`, and wall-clock is implemented in
wasi_poll.go, so the owner was reaching into a consumer for it.
Filesystem only needs the type because descriptor-stat carries three
timestamps.

Exporting it from there made that worse, since it put a clocks type on a
filesystem file's public surface, and left wasi_otel importing the whole
wasip2 implementation -- filesystem, sockets, http -- to obtain two
integer fields it needs for spans, logs and metrics. Nothing about
wasi:otel involves WASI.

imports/wasip2/wit declares the types instead. It depends only on the
component package and holds no implementation, so a host implementing an
unrelated WIT interface can take a type from it without depending on
WASI. wasi_otel now imports that leaf rather than the implementation.

Types are declared as a TypeDesc, with a convenience that interns one
into a component.TypeTable. The descriptor form is what lets wasip2's own
private table intern the same record, so there is one definition of the
field list rather than one per table type.
