+++ imports/wasi_otel/wit.go
@@ -2,7 +2,7 @@ package wasi_otel
 
 import (
 	"github.com/samyfodil/wazy/component"
-	"github.com/samyfodil/wazy/imports/wasip2"
+	"github.com/samyfodil/wazy/imports/wasip2/wit"
 )
 
 // The WIT type graph, interned into a component.TypeTable.
@@ -46,7 +46,7 @@ func newCommonTypes() *commonTypes {
 	// wasi:clocks/wall-clock's datetime, which the otel WIT uses rather than
 	// defining a time type of its own -- so it is taken from the package that
 	// already declares it rather than written out again here.
-	t.datetime = wasip2.DatetimeType(t.tbl)
+	t.datetime = wit.DatetimeType(t.tbl)
 
 	t.keyValue = t.tbl.Record(
 		"key", component.Prim("string"),
