+++ imports/wasip2/wasi_fs.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/samyfodil/wazy"
 	"github.com/samyfodil/wazy/component"
+	"github.com/samyfodil/wazy/imports/wasip2/wit"
 	"github.com/samyfodil/wazy/sys"
 )
 
@@ -1896,36 +1897,16 @@ func wasiErrorCodeType(tbl *typeTable) component.TypeRef {
 	}})
 }
 
-// DatetimeDesc returns wasi:clocks/wall-clock's `datetime` record
-// (`record datetime { seconds: u64, nanoseconds: u32 }`).
+// wasiDatetimeType interns wasi:clocks/wall-clock's `datetime` into tbl.
 //
-// It is exported because datetime is not this package's type: any host
-// implementing a WIT interface that reuses it -- wasi:otel's spans and logs,
-// for instance -- needs the same record, and two hand-written copies of one
-// wire format is one copy too many. Use [DatetimeType] with a
-// component.TypeTable, or hand this to a table of your own.
-//
-// A datetime value is a record value of the two fields in order:
-// []component.Value{uint64(seconds), uint32(nanoseconds)}.
-func DatetimeDesc() component.RecordDesc {
-	return component.RecordDesc{Fields: []component.RecordField{
-		{Name: "seconds", Type: component.TypeRef{Primitive: "u64"}},
-		{Name: "nanoseconds", Type: component.TypeRef{Primitive: "u32"}},
-	}}
-}
-
-// DatetimeType interns [DatetimeDesc] into tbl and returns its TypeRef.
-func DatetimeType(tbl *component.TypeTable) component.TypeRef {
-	return tbl.Add(DatetimeDesc())
-}
-
-// wasiDatetimeType interns the same record into this package's own table.
-// This package never constructs a datetime value (descriptor-stat's three
-// timestamp fields are always `none` -- see stat's doc), but the type must
-// still resolve structurally for Flatten to compute descriptor-stat's joined
-// flat width, mirroring wasi.go's wasiStreamErrorType doc.
+// The record itself is declared in the wit package, since wall-clock owns it
+// and this file is only a consumer: filesystem needs it for descriptor-stat's
+// three timestamp fields. Those are always `none` (see stat's doc), so this
+// package never constructs a datetime value, but the type must still resolve
+// structurally for Flatten to compute descriptor-stat's joined flat width,
+// mirroring wasi.go's wasiStreamErrorType doc.
 func wasiDatetimeType(tbl *typeTable) component.TypeRef {
-	return tbl.add(DatetimeDesc())
+	return tbl.add(wit.DatetimeDesc())
 }
 
 // wasiDescriptorStatType interns wasi:filesystem/types' `descriptor-stat`
