+++ imports/wasip2/wasi_fs_test.go
@@ -990,60 +990,3 @@ func TestWasiFS_GetResources_NotInitialized(t *testing.T) {
 	_, err := fs.getResources()
 	requireErrContains(t, err, "resources handle table not yet initialized")
 }
-
-// TestDatetimeDesc pins the exported datetime record.
-//
-// It is public API that other host packages intern into their own tables
-// (wasi:otel's spans, logs and metrics all carry a datetime), and a record's
-// field names never reach the wire -- only order and types do. So a reordering
-// here would silently change the layout of every interface that reuses it, and
-// no guest test elsewhere would necessarily say why.
-func TestDatetimeDesc(t *testing.T) {
-	d := DatetimeDesc()
-
-	if len(d.Fields) != 2 {
-		t.Fatalf("expected 2 fields, got %d", len(d.Fields))
-	}
-	for i, want := range []struct{ name, prim string }{
-		{"seconds", "u64"},
-		{"nanoseconds", "u32"},
-	} {
-		if got := d.Fields[i].Name; got != want.name {
-			t.Errorf("field %d name = %q, want %q", i, got, want.name)
-		}
-		if got := d.Fields[i].Type.Primitive; got != want.prim {
-			t.Errorf("field %d type = %q, want %q", i, got, want.prim)
-		}
-	}
-
-	// DatetimeType interns that same record and returns a ref to it.
-	tbl := component.NewTypeTable()
-	ref := DatetimeType(tbl)
-	if ref.TypeIndex == nil {
-		t.Fatal("DatetimeType should return a table ref, not an inline primitive")
-	}
-	got, ok := tbl.Resolver()(*ref.TypeIndex).(component.RecordDesc)
-	if !ok {
-		t.Fatalf("interned entry = %T, want component.RecordDesc", tbl.Resolver()(*ref.TypeIndex))
-	}
-	if len(got.Fields) != len(d.Fields) {
-		t.Fatalf("interned record has %d fields, want %d", len(got.Fields), len(d.Fields))
-	}
-
-	// This package's own table interns the identical shape, so the two entry
-	// points cannot drift apart.
-	priv := &typeTable{}
-	privRef := wasiDatetimeType(priv)
-	if privRef.TypeIndex == nil {
-		t.Fatal("wasiDatetimeType should return a table ref")
-	}
-	privDesc, ok := priv.entries[*privRef.TypeIndex].(component.RecordDesc)
-	if !ok {
-		t.Fatalf("private entry = %T, want component.RecordDesc", priv.entries[*privRef.TypeIndex])
-	}
-	for i := range d.Fields {
-		if privDesc.Fields[i] != d.Fields[i] {
-			t.Errorf("private table field %d = %+v, want %+v", i, privDesc.Fields[i], d.Fields[i])
-		}
-	}
-}
